refactor: move folders to src
