+++ src/stories/1-Input.stories.tsx
@@ -4,8 +4,8 @@
  * File Created: Wednesday, 8th July 2020 1:55:18 am
  * Author: Gabriel Ulloa (gabriel@inventures.cl)
  * -----
- * Last Modified: Wednesday, 26th August 2020 7:13:39 pm
- * Modified By: Esperanza Horn (esperanza@inventures.cl)
+ * Last Modified: Tuesday, 1st September 2020 2:42:19 pm
+ * Modified By: Gabriel Ulloa (gabriel@inventures.cl)
  * -----
  * Copyright 2019 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
  * Terms and conditions defined in license.txt
@@ -24,8 +24,8 @@ import {
   RutValidator,
   LengthValidator,
   NumericValidator,
-} from '../hooks/validators';
-import { AccentRemoverFormatter, RutFormatter } from '../hooks/formatters';
+} from '../validators';
+import { AccentRemoverFormatter, RutFormatter } from '../formatters';
 import Select from '@material-ui/core/Select';
 import FormControl from '@material-ui/core/FormControl';
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
@@ -65,7 +65,7 @@ export const InputForRut = () => {
   const [value, setValue, status, errors, handleBlur] = useInput('', {
     formatter: new RutFormatter(),
     validators: [
-      validLength && new LengthValidator(validLength, {min: 11, max: 12}),
+      validLength && new LengthValidator(validLength, { min: 11, max: 12 }),
       required && new RequiredValidator(required),
       incomplete && new RutFormatValidator(incomplete),
       valid && new RutValidator(valid),
@@ -178,13 +178,7 @@ export const InputForPhone = () => {
   );
   const debounceTime = number('Debounce time (ms)', 800);
 
-  const [
-    value,
-    setValue,
-    status,
-    errors,
-    handleBlur,
-  ] = useInput('', {
+  const [value, setValue, status, errors, handleBlur] = useInput('', {
     validators: [
       incompleteNumber &&
         new LengthValidator(incompleteNumber, country.countryDigitLength),
